break into beamparticle-core and remove stuff
This is a breaking change and incomplete. There is
a lot of work which is required to remove eleveldb and
instead use the filesystem of gitbackend to read and
write functions. This is largely to remove hard
dependency on eleveldb, which must give way for
simplistic mechanism. There is a local git repository
already, so lets use that for now before using
egraphdb as well.